Split dxf builder helper tests
